requirements.txt
asgiref
Django~=5.1.1
django-cors-headers
djangorestframework~=3.15.2
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
requests~=2.32.3